Code Cleanup
