+++ Assets/Resources/Systems.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8b0284c4eee09a1459954e8b407228d4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
