+++ Assets/Resources/Systems/StatSystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5bdc9b63fac5cef4e8a9268fc32463e1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
