+++ Assets/Resources/Systems/StatSystem/Database.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 73af46b61027d8a49b05c1ecf6e4830b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
