+++ Assets/Resources/Systems/StatSystem/Database/StatTypeDatabase.asset
@@ -1,0 +1,14 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1d31c9852da5c0b4f9b349d1e5887e01, type: 3}
+  m_Name: StatTypeDatabase
+  m_EditorClassIdentifier: 
+  _objects: []
