+++ Assets/Resources/Systems/StatSystem/Database/StatTypeDatabase.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2b9ddf4a029ffea44af0931ca36bb6cd
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
