--- Assets/Scripts/Viral/StatSystem/StatCollections/EnemyStatCollection.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 4933b8a117216254aaa0817d84bb8406
-timeCreated: 1483402380
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
